Tests: Rename jasmine-server-tests to server-tests.
