+++ Gruntfile.js
@@ -8,7 +8,7 @@ module.exports = function(grunt) {
 	var fs = require('fs');
 	var usePhantom = process.env.TRAVIS_PULL_REQUEST != 'false' || process.env.BROWSER == 'phantomjs';
 	var distTasks = JSON.parse(fs.readFileSync('Tests/dist-tasks.json'));
-	var serverSpecsJasmine = require('./Tests/jasmine-server-tests');
+	var serverTests = require('./Tests/server-tests');
 	var options = require('./Tests/gruntfile-options');
 
 	grunt.initConfig({
@@ -57,17 +57,17 @@ module.exports = function(grunt) {
 	var serverBuild = ['clean:specs', 'packager:server', 'packager:specs-server'];
 
 	var tasks = options.travis.build == 'default' ? compatBuild : options.travis.build == 'server' ? serverBuild : nocompatBuild;
-	tasks = options.travis.build == 'server' ? tasks.concat('server-jasmine-runner') : usePhantom ? tasks.concat('karma:continuous') : tasks.concat('karma:sauceTask');
+	tasks = options.travis.build == 'server' ? tasks.concat('server-spec-runner') : usePhantom ? tasks.concat('karma:continuous') : tasks.concat('karma:sauceTask');
 
 	// to run only server specs and avoid running browser specs for in the different versions of node and iojs
 	if (options.travis.build && options.travis.build != 'server' && process.version.slice(1, 5) != '0.10') tasks = [];
 
 	grunt.registerTask('default', compatBuild.concat('karma:continuous'));		// local testing - compat build
 	grunt.registerTask('nocompat', nocompatBuild.concat('karma:continuous'));	// local testing - no compat build
-	grunt.registerTask('server', serverBuild.concat('server-jasmine-runner'));	// local testing - server build
-	grunt.registerTask('server-jasmine-runner', function(){						// Travis server specs
+	grunt.registerTask('server', serverBuild.concat('server-spec-runner'));	// local testing - server build
+	grunt.registerTask('server-spec-runner', function(){						// Travis server specs
 		var done = this.async();
-		serverSpecsJasmine(done);
+		serverTests(done);
 	});
 	grunt.registerTask('default:travis', tasks);								// Travis & Sauce Labs
 	grunt.registerTask('distBuild', [											// task to build and test /dist files
